Uplift `Predicate::allow_normalization` to `rustc_type_ir`
